#158 refactor : RateBadge 관련 리팩터링
- 파일명 DeltaBadge -> RateBadge로 변경
- 예상 연 배당률 퍼센트에 화살표 아이콘 제거
