+++ src/components/Watchlist/WatchlistTable/WatchlistTableRow.tsx
@@ -1,6 +1,6 @@
 import { WatchlistItemType } from "@api/watchlist";
 import useWatchlistItemDeleteMutation from "@api/watchlist/queries/useWatchlistItemDeleteMutation";
-import RateBadge from "@components/common/Badges/DeltaBadge";
+import RateBadge from "@components/common/Badges/RateBadge";
 import CheckBox from "@components/common/Checkbox/Checkbox";
 import { CustomTooltip } from "@components/common/CustomTooltip";
 import { Icon } from "@components/common/Icon";
