+++ src/pages/StockPage.tsx
@@ -1,7 +1,7 @@
 import useStockItemQuery from "@api/stock/queries/useStockPageQeury";
 import AlertDropdown from "@components/Stock/TargetPriceAlertDropdown/TargetPriceAlertDropdown";
 import { WatchlistHasStockDropdown } from "@components/Stock/WatchlistHasStock/WatchlistHasStockDropdown";
-import RateBadge from "@components/common/Badges/DeltaBadge";
+import RateBadge from "@components/common/Badges/RateBadge";
 import Header from "@components/common/Header/Header";
 import designSystem from "@styles/designSystem";
 import { thousandsDelimiter } from "@utils/delimiters";
